refactor: Move kissutil from src into cmd/samoyed-kissutil
kissutil.go was tightly coupled to package direwolf internals (ax25_*,
kiss_* helpers and types) rather than only its public API, so moving it
required exporting thin wrappers for what it needs (src/kissutil_api.go)
and adding an exported callback hook (KissutilKissProcessMsg) so
kiss_frame.go's KISSUTIL branch can reach kissutil's message handling
without src depending on cmd/samoyed-kissutil.

The exported wrapper aliases/functions use ALevel, KISSFrame, and AX25*
naming to match Go initialism conventions, rather than the mixed styles
(AlevelT, KissFrameT, Ax25*) used in an earlier revision.
